backend/api-server/requirements.txt
fastapi
uvicorn
transformers
torch
scikit-learn
sqlite-utils
python-dotenv
openai
